Securing the views

+++ products/views.py
@@ -73,7 +73,10 @@ def quick_view(request, product_id):
 @login_required
 def add_product(request):
     """ Add a product to the store """
-    
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not a superuser.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -95,6 +98,11 @@ def add_product(request):
 @login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not a superuser.')
+        return redirect(reverse('home'))
+
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -119,6 +127,10 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     """ Delete product from store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you are not a superuser.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
